Restructure socket.io files
